backend/requirements.txt
Django>=5.1.0
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
requests>=2.31.0